% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crash_motorcycle}
\alias{crash_motorcycle}
\title{Motorcycle and Moped Crash Data in Florida}
\format{
A data frame with 32707 rows and 24 variables:

\describe{
\item{xid}{a unique identifier for each crash.}
\item{year}{calendar year of the crash as a string.}
\item{injsever}{injury severity is coded 1-6 with 1 None; 2 Possible; 3 Non-Incapacitating;
4 Incapacitating; 5 Fatal (within 30 days); and 6 Non-traffic Fatality}
\item{alcinvcd}{alcohol/drugs involved is coded 1-4}
\item{pop}{population as integer}
}
}
\source{
\url{https://gis-fdot.opendata.arcgis.com/datasets/fdot::motorcycle-moped-crashes/about}
}
\usage{
crash_motorcycle
}
\description{
A dataset containing motorcycle and moped crashes in Florida cities. The data was from
two sources: the Florida Department of Transportation (FDOT) and the U.S. Census Bureau.
This file shows crash locations for Long-Form-reported crashes within the state of Florida and
is subset to those occurring within towns and having a county-city code assigned.
}
\keyword{datasets}
