% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/florida_city_codes.R
\docType{data}
\name{fl_city_codes}
\alias{fl_city_codes}
\title{Florida city codes}
\format{
A data frame with 637 rows and 2 variables:
\describe{
\item{city_code}{Four digit unique city identifier}
\item{city}{a city in Florida}
}
}
\usage{
fl_city_codes
}
\description{
A dataset containing the city codes and city names for Florida.
}
\references{
Florida  Department of Highway Safety and Motor Vehicles,
Uniform Traffic Crash Report Manual, Appendix C (Rev'd 02/05/2019).
}
\keyword{datasets}
