% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/florida_city_pop.R
\docType{data}
\name{fl_city_pop}
\alias{fl_city_pop}
\title{Florida annual population by city 2010 - 2019}
\format{
A data frame with 10 rows and 3 variables:
\describe{
\item{city}{names of florida cities}
\item{year}{year as integer}
\item{pop}{number of persons}

}
}
\source{
\href{https://www.census.gov/data/tables/time-series/demo/popest/2010s-state-total.html}{US Census Bureau, State Population Totals and Components of Change: 2010-2019}
}
\usage{
fl_city_pop
}
\description{
A dataset containing the annual population for Florida cities from 2010 to 2019.
}
\keyword{datasets}
