% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fl_county_codes.R
\docType{data}
\name{fl_county_codes}
\alias{fl_county_codes}
\title{Florida county FIPS codes}
\format{
A data frame with 67 rows and 2 variables:
\describe{
\item{fips}{fips code}
\item{county}{county name}

}
}
\source{
\href{https://github.com/kjhealy/fips-codes}{Kieran Healy at Github}
}
\usage{
fl_county_codes
}
\description{
A dataset containing the five digit Federal Information Processing Codes (FIPS) codes for Florida counties.
}
\keyword{datasets}
