% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/florida_county_pop.R
\docType{data}
\name{fl_county_pop}
\alias{fl_county_pop}
\title{Florida annual population by county 2010 - 2019}
\format{
A data frame with 10 rows and 3 variables:
\describe{
\item{city}{names of florida counties}
\item{year}{year as integer}
\item{pop}{number of persons}

}
}
\source{
\href{https://www.census.gov/data/datasets/time-series/demo/popest/2010s-counties-total.html}{US Census Bureau, County Population Totals: 2010-2019}
}
\usage{
fl_county_pop
}
\description{
A dataset containing the annual population for Florida counties from 2010 to 2019.
}
\keyword{datasets}
