% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/florida_state_pop.R
\docType{data}
\name{fl_state_pop}
\alias{fl_state_pop}
\title{Florida annual population 2010 - 2019}
\format{
A data frame with 10 rows and 3 variables:
\describe{
\item{state}{Florida}
\item{calendar_year}{The year as integer}
\item{pop}{Number of persons}

}
}
\source{
\href{https://www.census.gov/data/tables/time-series/demo/popest/2010s-state-total.html}{State Population Totals and Components of Change: 2010-2019}
}
\usage{
fl_state_pop
}
\description{
A dataset containing the annual population for Florida from 2010 to 2019.
}
\keyword{datasets}
